% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{BlankIfNA}
\alias{BlankIfNA}
\title{Replace NAs in a numeric vector x with 0. This function has the same logic
as the zeroifnull function in SQL. NAIfZero() does replace zeros with NA.
BlankIfNA() and NAIfBlank() do the same, but for character vectors.
Copied directly from DescTools::BlankIfNA()}
\usage{
BlankIfNA(x, blank = "")
}
\arguments{
\item{x}{the vector x, whose NAs should be overwritten with 0s.}

\item{blank}{a character to be used for "blank". Default is an empty string ("").}
}
\description{
Replace NAs in a numeric vector x with 0. This function has the same logic
as the zeroifnull function in SQL. NAIfZero() does replace zeros with NA.
BlankIfNA() and NAIfBlank() do the same, but for character vectors.
Copied directly from DescTools::BlankIfNA()
}
