% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp4.R
\name{build_mp4}
\alias{build_mp4}
\title{Build xaringan slides as a mp4 video file.}
\usage{
build_mp4(
  input,
  output_file = NULL,
  density = "100x100",
  framerate = 1,
  complex_slides = FALSE,
  partial_slides = FALSE,
  delay = 1
)
}
\arguments{
\item{input}{Path to a Rmd file, html file, pdf file, or a url.
If the input is a url to xaringan slides on a website, you must provide the
full url ending in ".html".}

\item{output_file}{Name of the output mp4 file.}

\item{density}{Resolution of each slide in the video. Defaults to
`"100x100"`.}

\item{framerate}{Video framerate in frames per seconds. Defaults to 1}

\item{complex_slides}{For "complex" slides (e.g. slides with panelsets or
other html widgets or advanced features), set `complex_slides = TRUE`.
Defaults to `FALSE`. This will use the {chromote} package to iterate through
the slides at a pace set by the `delay` argument. Requires a local
installation of Chrome.}

\item{partial_slides}{Should partial (continuation) slides be
included in the output? If `FALSE`, the default, only the complete slide
is included in the PDF.}

\item{delay}{Seconds of delay between advancing to and printing
a new slide. Only used if `complex_slides = TRUE` or `partial_slides =
TRUE`.}
}
\description{
Build xaringan slides as a mp4 video file. The function builds to the pdf,
converts each slide in the pdf to a png, and then converts the deck of
png files to a mp4 video file.
}
\examples{
\dontrun{
# Build mp4 from Rmd, html, pdf, or url
build_mp4("slides.Rmd")
build_mp4("slides.html")
build_mp4("slides.pdf")
build_mp4("https://jhelvy.github.io/xaringanBuilder/reference/figures/slides.html")
}
}
