% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{build_pdf_complex}
\alias{build_pdf_complex}
\title{Print "complex" xaringan slides to PDF}
\usage{
build_pdf_complex(input, output_file, partial_slides, delay)
}
\arguments{
\item{input}{Path to Rmd or html file of xaringan slides.}

\item{output_file}{The name of the output file. If `NULL` (the default) then
the output filename will be based on filename for the input file.
If a filename is provided, a path to the output file can also be provided.}

\item{partial_slides}{Should partial (continuation) slides be
included in the output? If `FALSE`, the default, only the complete slide
is included in the PDF.}

\item{delay}{Seconds of delay between advancing to and printing
a new slide.}
}
\description{
Prints "complex" xaringan slides (e.g. slides with panelsets or other html
widgets or advanced features) to a PDF file. Requires a local installation
of Chrome.
}
