% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{build_pdf_simple}
\alias{build_pdf_simple}
\title{Print "simple" xaringan slides to PDF}
\usage{
build_pdf_simple(input, output_file)
}
\arguments{
\item{input}{Path to Rmd or html file of xaringan slides.}

\item{output_file}{The name of the output file. If `NULL` (the default) then
the output filename will be based on filename for the input file.
If a filename is provided, a path to the output file can also be provided.}
}
\description{
Prints "simple" xaringan slides - those without panelsets or other html
widgets or advanced features, and also slides without partial slides.
}
