% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{build_pptx}
\alias{build_pptx}
\title{Build xaringan slides as pptx file.}
\usage{
build_pptx(input, output_file = NULL, density = "72x72")
}
\arguments{
\item{input}{Path to Rmd, html, or pdf file of xaringan slides.}

\item{output_file}{Name of the output pptx file.}

\item{density}{Resolution of the resulting pptx file.}
}
\description{
Build xaringan slides as pptx file.
}
\examples{
\dontrun{
# Build gif from Rmd, html, or pdf file
build_pptx("slides.Rmd")
build_pptx("slides.html")
build_pptx("slides.pdf")
}
}
