% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thumbnail.R
\name{build_thumbnail}
\alias{build_thumbnail}
\title{Build png thumbnail image of first slide.}
\usage{
build_thumbnail(input, output_file = NULL)
}
\arguments{
\item{input}{Path to a Rmd file or html file / url of xaringan slides. If
the input is a url to xaringan slides on a website, you must provide the
full url ending in ".html".}

\item{output_file}{Name of the output png file.}
}
\description{
Build png thumbnail image of first xaringan slide. Requires a local
installation of Chrome.
}
\examples{
\dontrun{
# Build first slide thumbnail from Rmd or html file
build_thumbnail("slides.Rmd")
build_thumbnail("slides.html")
}
}
