% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-deprecated.R
\name{deprecated-build}
\alias{deprecated-build}
\alias{build_html}
\alias{build_pdf}
\alias{build_png}
\alias{build_gif}
\alias{build_mp4}
\alias{build_pptx}
\alias{build_social}
\title{Deprecated Build Functions}
\usage{
build_html(...)

build_pdf(...)

build_png(...)

build_gif(...)

build_mp4(...)

build_pptx(...)

build_social(...)
}
\arguments{
\item{...}{Parameters passed to the new \verb{to_*()} function}
}
\value{
See the corresponding new function for the appropriate return value.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\pkg{renderthis}, under the name \pkg{xaringanBuilder}, previously provided
the same functionality using \code{build_} functions. To be consistent with the
new package name, these function names have also been changed.
\itemize{
\item \code{build_html()} is now \code{\link[=to_html]{to_html()}}
\item \code{build_pdf()} is now \code{\link[=to_pdf]{to_pdf()}}
\item \code{build_png()} is now \code{\link[=to_png]{to_png()}}
\item \code{build_gif()} is now \code{\link[=to_gif]{to_gif()}}
\item \code{build_mp4()} is now \code{\link[=to_mp4]{to_mp4()}}
\item \code{build_pptx()} is now \code{\link[=to_pptx]{to_pptx()}}
\item \code{build_social()} is now \code{\link[=to_social]{to_social()}}
}
}
\keyword{internal}
