% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_paths}
\alias{get_paths}
\title{Returns a named list of the split full path into its components.}
\usage{
get_paths(input)
}
\arguments{
\item{input}{Path to Rmd file of xaringan slides.}
}
\description{
Returns a named list of the split full path into its components.
}
