% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SplitPath}
\alias{SplitPath}
\title{Split a full path in its components. This is specifically an issue
in Windows and not really interesting for other OSs. Modified
from DescTools::SplitPath()}
\usage{
SplitPath(path, last.is.file = NULL)
}
\arguments{
\item{path}{a path}

\item{last.is.file}{logical, determining if the basename should be
interpreted as filename or as last directory. If set to NULL (default),
the last entry will be interpreted if the last character is either \ or
/ and as filename else.}
}
\description{
Split a full path in its components. This is specifically an issue
in Windows and not really interesting for other OSs. Modified
from DescTools::SplitPath()
}
