% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{StrExtract}
\alias{StrExtract}
\title{Extract a part of a string, defined as regular expression.
Copied directly from DescTools::StrExtract()}
\usage{
StrExtract(x, pattern, ...)
}
\arguments{
\item{x}{a character vector where matches are sought, or an object which
can be coerced by as.character to a character vector.}

\item{pattern}{character string containing a regular expression (or
character string for fixed = TRUE) to be matched in the given character
vector. Coerced by as.character to a character string if possible.
If a character vector of length 2 or more is supplied, the first
element is used with a warning. Missing values are not allowed.}

\item{...}{the dots are passed to the the internally used function
regexpr(), which allows to use e.g. Perl-like regular expressions.}
}
\description{
Extract a part of a string, defined as regular expression.
Copied directly from DescTools::StrExtract()
}
