% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{to_all}
\alias{to_all}
\title{Build xaringan slides to multiple outputs.}
\usage{
to_all(
  from,
  include = c("html", "social", "pdf", "png", "gif", "mp4", "pptx"),
  exclude = NULL,
  complex_slides = FALSE,
  partial_slides = FALSE,
  delay = 1,
  density = 100,
  slides = "all",
  fps = 1
)
}
\arguments{
\item{from}{Path to Rmd file of xaringan slides.}

\item{include}{A vector of the different output types to build. Options are
\code{"html"}, \code{"social"}, \code{"pdf"}, \code{"png"}, \code{"gif"}, \code{"mp4"}, and \code{"pptx"}.
Defaults to \code{c("html", "social", "pdf", "png", "gif", "mp4", "pptx")}.}

\item{exclude}{A vector of the different output types to NOT build. Options
are \code{"html"}, \code{"social"}, \code{"pdf"}, \code{"png"}, \code{"gif"}, \code{"mp4"}, and \code{"pptx"}.
Defaults to \code{NULL}, in which case all all output types are built.}

\item{complex_slides}{For "complex" slides (e.g. slides with panelsets or
other html widgets or advanced features), set \code{complex_slides = TRUE}.
Defaults to \code{FALSE}. This will use the {chromote} package to iterate
through the slides at a pace set by the \code{delay} argument. Requires a local
installation of Chrome.}

\item{partial_slides}{Should partial (continuation) slides be included in the
output? If \code{FALSE}, the default, only the complete slide is included in the
PDF.}

\item{delay}{Seconds of delay between advancing to and printing a new slide.
Only used if \code{complex_slides = TRUE} or \code{partial_slides = TRUE}.}

\item{density}{Resolution of the resulting pngs in each slide file. Defaults
to \code{100}.}

\item{slides}{A numeric or integer vector of the slide number(s) to render
as png files , or one of \code{"all"}, \code{"first"}, or \code{"last"}. Negative integers
select which slides \emph{not} to include. If more than one slide are included,
pngs will be returned as a zip file. Defaults to \code{"all"}, in which case
all slides are included.}

\item{fps}{Frames per second of the resulting mp4 file.}
}
\description{
Build xaringan slides to multiple outputs. Options are \code{"html"}
(\code{\link[=to_html]{to_html()}}), \code{"social"} (\code{\link[=to_social]{to_social()}}), \code{"pdf"} (\code{\link[=to_pdf]{to_pdf()}}), \code{"png"}
(\code{\link[=to_png]{to_png()}}), \code{"gif"} (\code{\link[=to_gif]{to_gif()}}), \code{"mp4"} (\code{\link[=to_mp4]{to_mp4()}}), and \code{"pptx"}
(\code{\link[=to_pptx]{to_pptx()}}).
}
\examples{
\dontrun{
# Builds every output by default
to_all("slides.Rmd")

# Both of these build html, pdf, and gif outputs
to_all("slides.Rmd", include = c("html", "pdf", "gif"))
to_all("slides.Rmd", exclude = c("social", "png", "mp4", "pptx"))
}

}
