% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handout.R
\name{to_handout}
\alias{to_handout}
\title{Render handout from slides}
\usage{
to_handout(
  from,
  to = NULL,
  include = c("preview", "content", "notes"),
  complex_slides = FALSE,
  partial_slides = FALSE,
  delay = 1,
  keep_intermediates = NULL,
  include_images = FALSE
)
}
\arguments{
\item{from}{Path to a Rmd file, html file, or a url. If \code{from}
is a url to slides on a website, you must provide the full url
ending in \code{".html"}.}

\item{to}{The directory where the slide handout, preview images, and
other slide data will be stored. Because many files are included in the
\code{to_handout()} output, a directory is required. If the directory
already exists it will be overwritten.}

\item{include}{A character vector of sections to include in the handout:
\itemize{
\item \code{"preview"}: a preview image of the slide.
\item \code{"content"}: the content extracted from the slide, discarding most
problematic content types, but including images if \code{include_images} is
\code{TRUE}.
\item \code{"notes"}: the presenter notes from the slide.
\item \code{"lines"}: a lined notes area.
}}

\item{complex_slides}{For "complex" slides (e.g. slides with panelsets or
other html widgets or advanced features), set \code{complex_slides = TRUE}.
Defaults to \code{FALSE}. This will use the {chromote} package to iterate
through the slides at a pace set by the \code{delay} argument. Requires a local
installation of Chrome.}

\item{partial_slides}{Should partial (continuation) slides be included in the
output? If \code{FALSE}, the default, only the complete slide is included in the
PDF.}

\item{delay}{Seconds of delay between advancing to and printing a new slide.
Only used if \code{complex_slides = TRUE} or \code{partial_slides = TRUE}.}

\item{keep_intermediates}{Should we keep the intermediate HTML file? Only
relevant if the \code{from} file is an \code{.Rmd} file. Default is \code{TRUE}
if the \code{to} file is written into the same directory as the \code{from} argument,
otherwise the intermediate file isn't kept.}

\item{include_images}{Should images be included in the slide \code{content}
section? The default is \code{FALSE} to keep mostly the text content of each
slide.}
}
\description{
Renders a presentation handout as an \code{.Rmd} and \code{.html} file from a
set of slides. The handout extracts a thumbnail preview for each slide in
the input, as well as the slide content and presenter notes. The final output
is a \code{.html} file that is suitable for publishing or printing. The output
also includes the \code{.Rmd} file used to create the handout so that you can
edit the content as needed and can re-render the handout without having to
call \code{to_handout()} again.
}
\examples{
\dontrun{

# Render handouts with image previews and lines
to_handout(
    from = "https://jhelvy.github.io/renderthis/reference/figures/slides.html",
    to = "handouts",
    include = c("preview", "lines"),
)

# Render handouts with image previews, content, and speaker notes
to_handout(
    from = "https://jhelvy.github.io/renderthis/reference/figures/slides.html",
    to = "handouts",
    include = c("preview", "content", "notes")
)
}

}
