% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaringan_to_pdf.R
\name{xaringan_to_pdf}
\alias{xaringan_to_pdf}
\title{Print xaringan slides to PDF}
\usage{
xaringan_to_pdf(
  input,
  output_file = NULL,
  delay = 1,
  include_partial_slides = FALSE
)
}
\arguments{
\item{input}{Path to Rmd or html file of xaringan slides.}

\item{output_file}{The name of the output file. If using NULL then
the output filename will be based on filename for the input file.
If a filename is provided, a path to the output file can also be provided.}

\item{delay}{Seconds of delay between advancing to and printing
a new slide.}

\item{include_partial_slides}{Should partial (continuation) slides be
  included in the output? If `FALSE`, the default, only the complete slide
is included in the PDF.}
}
\description{
Prints xaringan slides to a PDF file, even complicated slides
with panelsets or other html widgets or advanced features.
Requires a local installation of Chrome.
}
\examples{
\dontrun{
# Build pdf from Rmd or html file
xaringan_to_pdf("slides.Rmd")
xaringan_to_pdf("slides.html")
}
}
