% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ZeroIfNA}
\alias{ZeroIfNA}
\title{Replace NAs in a numeric vector x with 0. This function has the same
logic as the zeroifnull function in SQL. NAIfZero() does replace
zeros with NA. BlankIfNA() and NAIfBlank() do the same, but for
character vectors. Copied directly from DescTools::ZeroIfNA()}
\usage{
ZeroIfNA(x)
}
\arguments{
\item{x}{the vector x, whose NAs should be overwritten with 0s.}
}
\description{
Replace NAs in a numeric vector x with 0. This function has the same
logic as the zeroifnull function in SQL. NAIfZero() does replace
zeros with NA. BlankIfNA() and NAIfBlank() do the same, but for
character vectors. Copied directly from DescTools::ZeroIfNA()
}
