% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{build_all}
\alias{build_all}
\title{Build xaringan slides as multiple outputs, including html, pdf, gif, and thumbnail of first slide.}
\usage{
build_all(input, include = c("html", "pdf", "gif", "thumbnail"))
}
\arguments{
\item{input}{Path to Rmd file of xaringan slides.}

\item{include}{A vector of the different output types to build, including "html", "pdf", "gif", and "thumbnail".}
}
\description{
Build xaringan slides as multiple outputs, including html, pdf, gif, and thumbnail of first slide.
}
\examples{
\dontrun{
# Build all outputs from Rmd file
build_all("slides.Rmd", include = c("html", "pdf", "gif", "pptx", "thumbnail"))
}
}
