% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all.R
\name{build_all}
\alias{build_all}
\title{Build xaringan slides to multiple outputs.}
\usage{
build_all(
  input,
  include = c("html", "pdf", "gif", "pptx", "thumbnail", "social"),
  exclude = NULL,
  complex_slides = FALSE,
  partial_slides = FALSE,
  delay = 1
)
}
\arguments{
\item{input}{Path to Rmd file of xaringan slides.}

\item{include}{A vector of the different output types to build. Options are
`"html"`, `"pdf"`, `"gif"`, `"pptx"`, and `"thumbnail"` (a png image of the
first slide). Defaults to `c("html", "pdf", "gif", "pptx", "thumbnail")`.}

\item{exclude}{A vector of the different output types to NOT build. Options
are `"html"`, `"pdf"`, `"gif"`, `"pptx"`, and `"thumbnail"` (a png image of
the first slide). Defaults to `NULL`, in which case all all output types
are rendered.}

\item{complex_slides}{For "complex" slides (e.g. slides with panelsets or
other html widgets or advanced features), set `complex_slides = TRUE`.
Defaults to `FALSE`. This will use the {chromote} package to iterate through
the slides at a pace set by the `delay` argument. Requires a local
installation of Chrome.}

\item{partial_slides}{Should partial (continuation) slides be
included in the output? If `FALSE`, the default, only the complete slide
is included in the PDF.}

\item{delay}{Seconds of delay between advancing to and printing
a new slide. Only used if `complex_slides = TRUE` or `partial_slides =
TRUE`.}
}
\description{
Build xaringan slides to multiple outputs. Options are `"html"`, `"pdf"`,
`"gif"`, `"pptx"`, and `"thumbnail"` (a png image of the first slide).
}
\examples{
\dontrun{
# Builds every output by default
build_all("slides.Rmd")

# Choose which output types to include
build_all("slides.Rmd", include = c("html", "pdf", "gif"))

# Choose which output types to exclude
build_all("slides.Rmd", exclude = c("pptx", "thumbnail"))
}
}
