% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{build_gif}
\alias{build_gif}
\title{Build xaringan slides as gif file.}
\usage{
build_gif(input, output_file = NULL, density = "72x72", fps = 1)
}
\arguments{
\item{input}{Path to Rmd, html, or pdf file of xaringan slides.}

\item{output_file}{Name of the output gif file.}

\item{density}{Resolution of the resulting gif file.}

\item{fps}{Frames per second.}
}
\description{
Build xaringan slides as gif file.
}
\examples{
# Build gif from Rmd, html, or pdf file
build_gif(here::here("test", "slides.Rmd"))
build_gif(here::here("test", "slides.html"))
build_gif(here::here("test", "slides.pdf"))
}
