% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gif.R
\name{build_gif}
\alias{build_gif}
\title{Build xaringan slides as gif file.}
\usage{
build_gif(
  input,
  output_file = NULL,
  density = "72x72",
  fps = 1,
  complex_slides = FALSE,
  partial_slides = FALSE,
  delay = 1
)
}
\arguments{
\item{input}{Path to a Rmd file, html file, pdf file, or a url.
If the input is a url to xaringan slides on a website, you must provide the
full url ending in ".html".}

\item{output_file}{Name of the output gif file.}

\item{density}{Resolution of the resulting gif file. Defaults to
`"72x72"`.}

\item{fps}{Frames per second of the resulting gif file.}

\item{complex_slides}{For "complex" slides (e.g. slides with panelsets or
other html widgets or advanced features), set `complex_slides = TRUE`.
Defaults to `FALSE`. This will use the {chromote} package to iterate through
the slides at a pace set by the `delay` argument. Requires a local
installation of Chrome.}

\item{partial_slides}{Should partial (continuation) slides be
included in the output? If `FALSE`, the default, only the complete slide
is included in the PDF.}

\item{delay}{Seconds of delay between advancing to and printing
a new slide. Only used if `complex_slides = TRUE` or `partial_slides =
TRUE`.}
}
\description{
Build xaringan slides as a gif file. The function builds to the pdf and
then converts it to a gif.
}
\examples{
\dontrun{
# Build gif from Rmd, html, pdf, or url
build_gif("slides.Rmd")
build_gif("slides.html")
build_gif("slides.pdf")
build_gif("https://jhelvy.github.io/xaringanBuilder/reference/figures/slides.html")
}
}
