% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handout.R
\name{build_handout}
\alias{build_handout}
\title{Build handout from xaringan slides}
\usage{
build_handout(
  input,
  output_dir = NULL,
  include = c("preview", "content", "notes"),
  keep_intermediates = FALSE,
  partial_slides = FALSE,
  include_images = FALSE
)
}
\arguments{
\item{input}{Path to a Rmd file or html file / url of xaringan slides. If the
input is a url to xaringan slides on a website, you must provide the full
url ending in ".html".}

\item{output_dir}{The directory where the slide handout, preview images and
other slide data will be stored. Because many files are included in the
`build_handout()` output, a directory is required. If the directory exists
it will be overwritten.}

\item{include}{A character vector of sections to include in the handout:

  - `"preview"`: a preview image of the slide
  - `"content"`: the content extracted from the slide, discarding most
    problematic content types, but including images if `include_images` is
    `TRUE`.
  - `"notes"`: the presenter notes from the slide
  - `"lines"`: a lined notes area}

\item{keep_intermediates}{Should we keep the intermediate HTML file? Only
relevant if the `input` is an `.Rmd` file. Default is `TRUE` if the
`output_file` is written into the same directory as the `input`, otherwise
the intermediate file isn't kept.}

\item{partial_slides}{Should partial (continuation) slides be included in the
output? If `FALSE`, the default, only the complete slide is included in the
PDF.}

\item{include_images}{Should images be included in the slide `content`
section? The default is `FALSE` to keep mostly the text content of each
slide.}
}
\description{
Builds a presentation handout as an `.Rmd` and `.html` file from a set of
xaringan slides. The handout extracts a thumbnail preview for each slide in
the input, as well as the slide content and presenter notes. The final output
is an `.html` file that's suitable for publishing or printing. The output
also includes the `.Rmd` file used to create the handout so that you can
edit the content as needed and can re-render the handout without having to
call `build_handout()` again.
}
