% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{build_pdf}
\alias{build_pdf}
\title{Build xaringan slides as pdf file.}
\usage{
build_pdf(input, output_file = NULL)
}
\arguments{
\item{input}{Path to Rmd or html file of xaringan slides.}

\item{output_file}{Name of the output pdf file.}
}
\description{
Build xaringan slides as pdf file.
}
\examples{
\dontrun{
# Build pdf from Rmd or html file
build_pdf("slides.Rmd")
build_pdf("slides.html")
}
}
