% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{build_pdf}
\alias{build_pdf}
\title{Build xaringan slides as pdf file.}
\usage{
build_pdf(
  input,
  output_file = NULL,
  complex_slides = FALSE,
  partial_slides = FALSE,
  delay = 1,
  keep_intermediates = NULL
)
}
\arguments{
\item{input}{Path to a Rmd file or html file / url of xaringan slides. If the
input is a url to xaringan slides on a website, you must provide the full
url ending in ".html".}

\item{output_file}{The name of the output file. If `NULL` (the default) then
the output filename will be based on filename for the input file. If a
filename is provided, a path to the output file can also be provided.}

\item{complex_slides}{For "complex" slides (e.g. slides with panelsets or
other html widgets or advanced features), set `complex_slides = TRUE`.
Defaults to `FALSE`. This will use the {chromote} package to iterate
through the slides at a pace set by the `delay` argument. Requires a local
installation of Chrome.}

\item{partial_slides}{Should partial (continuation) slides be included in the
output? If `FALSE`, the default, only the complete slide is included in the
PDF.}

\item{delay}{Seconds of delay between advancing to and printing a new slide.
Only used if `complex_slides = TRUE` or `partial_slides = TRUE`.}

\item{keep_intermediates}{Should we keep the intermediate HTML file? Only
relevant if the `input` is an `.Rmd` file. Default is `TRUE` if the
`output_file` is written into the same directory as the `input`, otherwise
the intermediate file isn't kept.}
}
\description{
Build xaringan slides as a pdf file. Requires a local installation of Chrome.
If you set `complex_slides = TRUE` or `partial_slides = TRUE`, you will also
need to install the {chromote} and {pdftools} packages.
}
\examples{
\dontrun{
# Build pdf from Rmd, html, or url
build_pdf("slides.Rmd")
build_pdf("slides.html")
build_pdf("https://jhelvy.github.io/xaringanBuilder/reference/figures/slides.html")

# Build a pdf with partial (continuation) slides
build_pdf("slides.Rmd", partial_slides = TRUE)
build_pdf("slides.html", partial_slides = TRUE)

# Build a pdf of "complex" xaringan slides
build_pdf("slides_complex.Rmd", complex_slides = TRUE)
build_pdf("slides_complex.html", complex_slides = TRUE)

# Build a pdf of "complex" xaringan slides and include partial
# (continuation) slides
build_pdf(input = "slides_complex.Rmd",
          output_file = "slides_complex_partial.pdf",
          complex_slides = TRUE,
          partial_slides = TRUE)
build_pdf(input = "slides_complex.html",
          output_file = "slides_complex_partial.pdf",
          complex_slides = TRUE,
          partial_slides = TRUE)
}

}
