% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{build_pptx}
\alias{build_pptx}
\title{Build xaringan slides as pptx file.}
\usage{
build_pptx(input, output_file = NULL, density = "72x72")
}
\arguments{
\item{input}{Path to Rmd, html, or pdf file of xaringan slides.}

\item{output_file}{Name of the output pptx file.}

\item{density}{Resolution of the resulting pptx file.}
}
\description{
Build xaringan slides as pptx file. Note: If you want to include "complex"
slides (e.g. slides with panelsets or other html widgets or advanced
features), or you want to include partial (continuation) slides, then use
`build_pdf()` with `complex_slides = TRUE` and / or `partial_slides = TRUE`
to build the pdf first, then use build_gif() with the pdf as the input.
}
\examples{
\dontrun{
# Build gif from Rmd, html, or pdf file
build_pptx("slides.Rmd")
build_pptx("slides.html")
build_pptx("slides.pdf")
}
}
