% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx.R
\name{build_pptx}
\alias{build_pptx}
\title{Build xaringan slides as pptx file.}
\usage{
build_pptx(
  input,
  output_file = NULL,
  density = "100x100",
  complex_slides = FALSE,
  partial_slides = FALSE,
  delay = 1
)
}
\arguments{
\item{input}{Path to a Rmd file, html file, pdf file, or a url.
If the input is a url to xaringan slides on a website, you must provide the
full url ending in ".html".}

\item{output_file}{Name of the output pptx file.}

\item{density}{Resolution of the resulting pngs in each slide file.
Defaults to `"100x100"`.}

\item{complex_slides}{For "complex" slides (e.g. slides with panelsets or
other html widgets or advanced features), set `complex_slides = TRUE`.
Defaults to `FALSE`. This will use the {chromote} package to iterate through
the slides at a pace set by the `delay` argument. Requires a local
installation of Chrome.}

\item{partial_slides}{Should partial (continuation) slides be
included in the output? If `FALSE`, the default, only the complete slide
is included in the PDF.}

\item{delay}{Seconds of delay between advancing to and printing
a new slide. Only used if `complex_slides = TRUE` or `partial_slides =
TRUE`.}
}
\description{
Build xaringan slides as a pptx file. The function builds to the pdf and
then converts it into png images that are inserted on each slide in the
pptx file.
}
\examples{
\dontrun{
# Build pptx from Rmd, html, pdf, or url
build_pptx("slides.Rmd")
build_pptx("slides.html")
build_pptx("slides.pdf")
build_pptx("https://jhelvy.github.io/xaringanBuilder/reference/figures/slides.html")
}
}
