% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{build_thumbnail}
\alias{build_thumbnail}
\title{Build png thumbnail image of first xaringan slide.}
\usage{
build_thumbnail(input, output_file = NULL)
}
\arguments{
\item{input}{Path to Rmd or html file of xaringan slides.}

\item{output_file}{Name of the output png file.}
}
\description{
Build png thumbnail image of first xaringan slide.
}
\examples{
# Build first slide thumbnail from Rmd or html file
build_thumbnail(here::here("test", "slides.Rmd"))
build_thumbnail(here::here("test", "slides.html"))
}
