% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gif.R
\name{to_gif}
\alias{to_gif}
\title{Build xaringan slides as a gif file.}
\usage{
to_gif(
  input,
  output_file = NULL,
  density = 100,
  slides = "all",
  fps = 1,
  complex_slides = FALSE,
  partial_slides = FALSE,
  delay = 1,
  keep_intermediates = FALSE
)
}
\arguments{
\item{input}{Path to a Rmd file, html file, pdf file, or a url. If the input
is a url to xaringan slides on a website, you must provide the full url
ending in ".html".}

\item{output_file}{Name of the output gif file.}

\item{density}{Resolution of the resulting pngs in each slide file. Defaults
to `100`.}

\item{slides}{A numeric or integer vector of the slide number(s) to include
in the gif, or one of `"all"`, `"first"`, or `"last"`. Negative integers
select which slides _not_ to include. Defaults to `"all"`, in which case
all slides are included.}

\item{fps}{Frames per second of the resulting gif file.}

\item{complex_slides}{For "complex" slides (e.g. slides with panelsets or
other html widgets or advanced features), set `complex_slides = TRUE`.
Defaults to `FALSE`. This will use the {chromote} package to iterate
through the slides at a pace set by the `delay` argument. Requires a local
installation of Chrome.}

\item{partial_slides}{Should partial (continuation) slides be included in the
output? If `FALSE`, the default, only the complete slide is included in the
PDF.}

\item{delay}{Seconds of delay between advancing to and printing a new slide.
Only used if `complex_slides = TRUE` or `partial_slides = TRUE`.}

\item{keep_intermediates}{Should we keep the intermediate files used to build
the final output? The default is `FALSE`.}
}
\description{
Build xaringan slides as a gif video file. The function builds to the pdf,
converts each slide in the pdf to a png, and then converts the deck of
png files to a gif file.
}
\examples{
\dontrun{
# Build gif from Rmd, html, pdf, or url
to_gif("slides.Rmd")
to_gif("slides.html")
to_gif("slides.pdf")
to_gif("https://jhelvy.github.io/xaringanBuilder/reference/figures/slides.html")
}

}
