% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{to_pdf}
\alias{to_pdf}
\title{Render slides as pdf file.}
\usage{
to_pdf(
  from,
  to = NULL,
  complex_slides = FALSE,
  partial_slides = FALSE,
  delay = 1,
  keep_intermediates = NULL
)
}
\arguments{
\item{from}{Path to a Rmd file, html file, or a url. If \code{from}
is a url to slides on a website, you must provide the full url
ending in \code{".html"}.}

\item{to}{The name of the output file. If \code{NULL} (the default) then
the output filename will be based on filename for the \code{from} file. If a
filename is provided, a path to the output file can also be provided.}

\item{complex_slides}{For "complex" slides (e.g. slides with panelsets or
other html widgets or advanced features), set \code{complex_slides = TRUE}.
Defaults to \code{FALSE}. This will use the {chromote} package to iterate
through the slides at a pace set by the \code{delay} argument. Requires a local
installation of Chrome.}

\item{partial_slides}{Should partial (continuation) slides be included in the
output? If \code{FALSE}, the default, only the complete slide is included in the
PDF.}

\item{delay}{Seconds of delay between advancing to and printing a new slide.
Only used if \code{complex_slides = TRUE} or \code{partial_slides = TRUE}.}

\item{keep_intermediates}{Should we keep the intermediate HTML file? Only
relevant if the \code{from} file is an \code{.Rmd} file. Default is \code{TRUE}
if the \code{to} file is written into the same directory as the \code{from} argument,
otherwise the intermediate file isn't kept.}
}
\description{
Render slides as a pdf file. Requires a local installation of Chrome.
If you set \code{complex_slides = TRUE} or \code{partial_slides = TRUE}, you will also
need to install the {chromote} and {pdftools} packages.
}
\examples{
\dontrun{
# Render pdf from Rmd, html, or url
to_pdf("slides.Rmd")
to_pdf("slides.html")
to_pdf("https://jhelvy.github.io/renderthis/reference/figures/slides.html")

# Render a pdf with partial (continuation) slides
to_pdf("slides.Rmd", partial_slides = TRUE)
to_pdf("slides.html", partial_slides = TRUE)

# Render a pdf of "complex" slides
to_pdf("slides_complex.Rmd", complex_slides = TRUE)
to_pdf("slides_complex.html", complex_slides = TRUE)

# Render a pdf of "complex" slides and include partial
# (continuation) slides
to_pdf(from = "slides_complex.Rmd",
       to = "slides_complex_partial.pdf",
       complex_slides = TRUE,
       partial_slides = TRUE)
to_pdf(from = "slides_complex.html",
       to = "slides_complex_partial.pdf",
       complex_slides = TRUE,
       partial_slides = TRUE)
}

}
