% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social.R
\name{to_social}
\alias{to_social}
\title{Build png image of first slide sized for social media sharing.}
\usage{
to_social(from, to = NULL)
}
\arguments{
\item{from}{Path to Rmd file of input media (e.g., xaringan slides).}

\item{to}{The name of the output file. If using NULL then
the output filename will be based on filename for the `from` file.
If a filename is provided, a path to the output file can also be provided.}
}
\description{
Build png image of first xaringan slide for sharing on social media.
Requires a local installation of Chrome as well as the {webshot2} package:
`remotes::install_github("rstudio/webshot2")`.
}
\examples{
\dontrun{
# Build png image of first xaringan slide from Rmd file
# sized for sharing on social media
to_social("slides.Rmd")
}

}
