% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{xaringan_to_pdf}
\alias{xaringan_to_pdf}
\title{Print xaringan slides to PDF}
\usage{
xaringan_to_pdf(
  url,
  path = paste0(basename(dirname(url)), ".pdf"),
  delay = 0.5
)
}
\arguments{
\item{url}{The url to the xaringan slides, either online or local.
If the slides are stored locally, you may need to add `file://`
before the complete file path.}

\item{path}{The path where the PDF file should be saved}

\item{delay}{Seconds of delay between advancing to and printing
a new slide.}
}
\description{
Prints xaringan slides to a PDF file, even complicated slides
with panelsets or other html widgets or advanced features.
Requires a local installation of Chrome.
}
